Move directory
src/component/icon -> src/icon
src/component -> src/content-twitter/component
